refactor: update ArchiveCardComponent import path and add normalization for archive file names
