+++ Planarian.Web/src/Modules/Account/Components/AccountSettingsComponent.tsx
@@ -14,7 +14,7 @@ import { EnabledFieldsComponent } from "./EnabledFieldsComponent";
 import { FeatureKey, FeatureSettingVm } from "../Models/FeatureSettingVm";
 import { AccountService } from "../Services/AccountService";
 import { AppContext } from "../../../Configuration/Context/AppContext";
-import { ArchiveCardComponent } from "./Archive/ArchiveCardComponent";
+import { ArchiveCardComponent } from "./ArchiveCardComponent";
 
 const AccountSettingsComponent = () => {
   const { useBreakpoint } = Grid;
